requirements.txt
azure-functions
msrestazure
pyodbc
requests
